Move RDataFrame utilities into utils folder
