+++ CMakeLists.txt
@@ -103,9 +103,8 @@ install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/LICENSE
               DESTINATION ${CMAKE_INSTALL_DOCDIR})
 
 add_subdirectory(edm4hep)
-add_subdirectory(test)
-add_subdirectory(dataframe)
 add_subdirectory(utils)
+add_subdirectory(test)
 
 #--- create uninstall target ---------------------------------------------------
 include(cmake/EDM4HEPUninstall.cmake)
