+++ utils/src/dataframe.cc
@@ -1,4 +1,4 @@
-#include "edm4hepRDF.h"
+#include "edm4hep/utils/dataframe.h"
 #include "Math/Vector4D.h"
 
 namespace edm4hep {
